refactor: consolidate storage utilities
